Refactoring Registation and Login (servlet) added validation after sign_up
