+++ src/main/java/com/fruitella/todo/DAO/TodoUserDao.java
@@ -84,4 +84,3 @@ public class TodoUserDao implements UserDao {
     }
 
 }
-
