+++ src/main/java/com/fruitella/todo/connection/TodoSessionFactory.java
@@ -25,4 +25,3 @@ public class TodoSessionFactory {
     }
 
 }
-
