+++ src/main/java/com/fruitella/todo/entity/User.java
@@ -27,5 +27,4 @@ public class User {
     @Column(name = "email", length = 75, nullable = false)
     private String email;
 
-
 }
