+++ src/main/webapp/invalid.jsp
@@ -1,0 +1,19 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Alex
+  Date: 03.03.2023
+  Time: 15:16
+  To change this template use File | Settings | File Templates.
+--%>
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <title>Login invalid</title>
+</head>
+<body>
+<h1>Error login</h1>
+<p>Invalid username or password ${loginBean.errorMessage}</p>
+</body>
+</html>
+
