+++ src/main/webapp/valid.jsp
@@ -1,0 +1,19 @@
+<%--
+  Created by IntelliJ IDEA.
+  User: Alex
+  Date: 03.03.2023
+  Time: 15:15
+  To change this template use File | Settings | File Templates.
+--%>
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+  <meta charset="UTF-8">
+  <title>Login Successful</title>
+</head>
+<body>
+<h1>Login successful</h1>
+<p>Welcome to TODO application! ${loginBean.username}</p>
+</body>
+</html>
+
